# openapi.yml
openapi: 3.0.1
info:
  title: mitum-currency - digest API
  description: >-
    This is API description of digest service of mitum-currency.
  contact:
    name: Spike^ekipS
    url: https://github.com/ProtoconNet/mitum-currency
    email: spikeekips@gmail.com
  license:
    name: GPL-3.0
    url: https://github.com/ProtoconNet/mitum-currency/blob/master/LICENSE
  version: 0.0.1
...
servers:
- url: https://localhost:54322
...
paths:
...
  /block/operation/{operation_fact_hash}:
    get:
      tags:
      - block
      summary: Operation by it's fact hash.
      description: >-
        operation by it's **fact** *hash*, not it's own *hash*.
      operationId: operation
      parameters:
        - name: operation_fact_hash
          in: path
          description: >-
              *fact* *hash* of operation.
          required: true
          schema:
            type: string
      responses:
        500:
          description: problems in processing.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        404:
          description: not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        200:
          description: hal document of operation
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/OperationHAL'
          headers:
            X-Rate-Limit:
              description: calls per hour allowed by the user
              schema:
                type: integer
                format: int32
            X-Rate-Remaining:
              description: remains request count
              schema:
                type: integer
                format: int32
            X-Rate-Reset:
              description: timestamp to reset limit
              schema:
                type: integer
                format: int64
...
  /block/operations:
    get:
      tags:
      - block
      summary: All the operations
      operationId: operations
      parameters:
        - name: offset
          in: query
          schema:
            type: string
            example: "2,0"
          description: >-
            *operation*s after *offset*.
        - name: reverse
          in: query
          schema:
            type: boolean
            example: false
            default: false
          description: >-
            *operation*s by reverse order.
      responses:
        500:
          description: problems in processing.
          content:
            application/problem+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Problem'
                  - type: object
                    properties:
                      title:
                        type: string
                        example: "...."
                      detail:
                        type: string
                        example: "...."
        404:
          description: no more operations
          content:
            application/problem+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Problem'
                  - type: object
                    properties:
                      title:
                        type: string
                        example: "operations not found"
                      detail:
                        type: string
                        example: "...."
        200:
          description: hal document of operations
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/OperationsHAL'
          headers:
            X-Rate-Limit:
              description: calls per hour allowed by the user
              schema:
                type: integer
                format: int32
            X-Rate-Remaining:
              description: remains request count
              schema:
                type: integer
                format: int32
            X-Rate-Reset:
              description: timestamp to reset limit
              schema:
                type: integer
                format: int64
...
  /currency:
    get:
      tags:
      - currency
      summary: Currencies information
      operationId: currencies
      responses:
        500:
          description: problems in processing.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        200:
          description: hal document of currencies
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/CurrenciesHAL'
          headers:
            X-Rate-Limit:
              description: calls per hour allowed by the user
              schema:
                type: integer
                format: int32
            X-Rate-Remaining:
              description: remains request count
              schema:
                type: integer
                format: int32
            X-Rate-Reset:
              description: timestamp to reset limit
              schema:
                type: integer
                format: int64
  /currency/{currency_id}:
    get:
      tags:
      - currency
      summary: Currency information
      operationId: currency
      parameters:
        - name: currency_id
          in: path
          description: currency unique id(or name)
          required: true
          schema:
            $ref: '#/components/schemas/CurrencyID'
      responses:
        500:
          description: problems in processing.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        200:
          description: hal document of currency of *currency_id*
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/CurrencyHAL'
          headers:
            X-Rate-Limit:
              description: calls per hour allowed by the user
              schema:
                type: integer
                format: int32
            X-Rate-Remaining:
              description: remains request count
              schema:
                type: integer
                format: int32
            X-Rate-Reset:
              description: timestamp to reset limit
              schema:
                type: integer
                format: int64
components:
  schemas:
    Hint:
      type: string
      example: a015:0.0.1
    Problem:
      type: object
      required:
      - _hint
      - title
      - type
      - detail
      properties:
        _hint:
          allOf:
            - $ref: '#/components/schemas/Hint'
            - type: string
              example: a017:0.0.1
              default: a017:0.0.1
        title:
          type: string
          example: "failed to sign"
        type:
          type: string
          example: "https://github.com/ProtoconNet/mitum-currency/problems/others"
        detail:
          type: string
          example: "invalid privatekey"
...
    HAL:
      type: object
      required:
      - _hint
      - _links
      properties:
        _hint:
          allOf:
            - $ref: '#/components/schemas/Hint'
            - type: string
              example: a016:0.0.1
              default: a016:0.0.1
        _embedded:
          type: object
          default: {}
        _links:
          type: object
          properties:
            self:
              allOf:
                - $ref: '#/components/schemas/HALLink'
                - type: object
                  properties:
                    href:
                      type: string
                      default: /
                      example: /
                    templated:
                      type: boolean
                      default: false
                      example: false
    HALLink:
      type: object
      required:
        - href
      properties:
        href:
          type: string
        templated:
          type: boolean
          description: The link is is templated link, not used without it's arguments.
          default: false
          example: false
...
    Height:
      description: >-
        height of *block*.
      type: integer
      format: int64
...
    BaseFact:
      type: object
      required:
      - _hint
      - hash
      - token
      properties:
        _hint:
          allOf:
            - $ref: '#/components/schemas/Hint'
            - type: string
              default: a005:0.0.1
              example: a005:0.0.1
        hash:
          type: string
          format: hash
          example: 4jhzcudKgtoPGR6rA7Fuxmfwz3C8KGiP5MEKXuBXcW9j
        token:
          type: string
          format: bytes
          example: cmFpc2VkIGJ5
    Operation:
      type: object
      required:
      - _hint
      - fact_signs
      - hash
      - fact
      properties:
        _hint:
          allOf:
            - $ref: '#/components/schemas/Hint'
            - type: string
              default: a008:0.0.1
              example: a008:0.0.1
        fact_signs:
          type: array
          items:
            $ref: '#/components/schemas/FactSign'
        hash:
          description: operation hash
          type: string
          format: hash
          example: 8u4ua32jnYjS1BP8jBoL2BYqSwdZTno8vd4NzujbZQbH
        fact:
          type: object
          required:
          - _hint
          - hash
          - token
          properties:
            _hint:
              allOf:
                - $ref: '#/components/schemas/Hint'
                - type: string
                  default: a008:0.0.1
                  example: a008:0.0.1
            hash:
              description: operation fact hash
              type: string
              format: hash
              example: 9u4ua32jnYjS1BP8jBoL2BYqSwdZTno8vd4NzujbZQbH
            token:
              type: string
              format: bytes
              example: cmFpc2VkIGJ5
        memo:
          type: string
          example: "show me and find me"
    OperationHAL:
      allOf:
        - $ref: '#/components/schemas/HAL'
        - type: object
          properties:
            hint:
              type: object
              properties:
                name:
                  type: string
                  default: mitum-currency-operation-value
                  example: mitum-currency-operation-value
                hint:
                  type: string
                  default: a019:0.0.1
                  example: a019:0.0.1
            _embedded:
              $ref: '#/components/schemas/OperationValue'
            _links:
              type: object
              properties:
                self:
                  allOf:
                    - $ref: '#/components/schemas/HALLink'
                    - type: object
                      properties:
                        href:
                          type: string
                          example: /operation/6GymxmuvhgUfAKWYQUKKveYh1H8b981vQaZcMsbWqykS
                operation:{hash}:
                  description: request `/operation/{hash}`
                  allOf:
                    - $ref: '#/components/schemas/HALLink'
                    - type: object
                      properties:
                        href:
                          type: string
                          default: /operation/{hash:(?i)[0-9a-z][0-9a-z]+}
                          example: /operation/{hash:(?i)[0-9a-z][0-9a-z]+}
                        templated:
                          type: boolean
                          default: true
                          example: true
                block:
                  description: request `/block/{height}` of the operation
                  allOf:
                    - $ref: '#/components/schemas/HALLink'
                    - type: object
                      properties:
                        href:
                          type: string
                          example: /block/244
                manifest:
                  description: request `/block/{height}/manifest` of the operation
                  allOf:
                    - $ref: '#/components/schemas/HALLink'
                    - type: object
                      properties:
                        href:
                          type: string
                          example: /block/244/manifest
...
    OperationsHAL:
      allOf:
        - $ref: '#/components/schemas/HAL'
        - type: object
          properties:
            _embedded:
              type: array
              items:
                $ref: '#/components/schemas/OperationHAL'
            _links:
              type: object
              properties:
                self:
                  allOf:
                    - $ref: '#/components/schemas/HALLink'
                    - type: object
                      properties:
                        href:
                          type: string
                          example: /block/operations
                next:
                  description: >-
                    next operations with *offset*.
                  allOf:
                    - $ref: '#/components/schemas/HALLink'
                    - type: object
                      properties:
                        href:
                          type: string
                          example: /block/operations?offset=2,0
                reverse:
                  description: >-
                    operations by reverse oder of self.
                  allOf:
                    - $ref: '#/components/schemas/HALLink'
                    - type: object
                      properties:
                        href:
                          type: string
                          example: /block/operations?reverse=1
...
    CreateAccounts:
      allOf:
        - $ref: '#/components/schemas/Operation'
        - type: object
          properties:
            fact:
              $ref: '#/components/schemas/CreateAccountsFact'
    KeyUpdater:
      allOf:
        - $ref: '#/components/schemas/Operation'
        - type: object
          properties:
            fact:
              $ref: '#/components/schemas/KeyUpdaterFact'
    Transfers:
      allOf:
        - $ref: '#/components/schemas/Operation'
        - type: object
          properties:
            fact:
              $ref: '#/components/schemas/TransfersFact'
    CurrencyRegister:
      allOf:
        - $ref: '#/components/schemas/Operation'
        - type: object
          properties:
            fact:
              $ref: '#/components/schemas/CurrencyRegisterFact'
    CurrencyPolicyUpdater:
      allOf:
        - $ref: '#/components/schemas/Operation'
        - type: object
          properties:
            fact:
              $ref: '#/components/schemas/CurrencyPolicyUpdaterFact'
    CreateAccountsFact:
      allOf:
        - $ref: '#/components/schemas/BaseFact'
        - type: object
          required:
          - sender
          - items
          properties:
            _hint:
              allOf:
                - $ref: '#/components/schemas/Hint'
                - type: string
                  example: a005:0.0.1
                  default: a005:0.0.1
            hash:
              description: >-
                The value of hash will be generated automatically by builder. *Don't need to edit*.
              type: string
              format: hash
              example: 4jhzcudKgtoPGR6rA7Fuxmfwz3C8KGiP5MEKXuBXcW9j
            token:
              description: >-
                Replace your own token. *token* value should be encoded by base64.
              type: string
              format: bytes
              example: cmFpc2VkIGJ5
            sender:
              allOf:
                - $ref: '#/components/schemas/AccountAddress'
                - description: Replace your own sender address.
            items:
              type: array
              items:
                type: object
                required:
                - keys
                - amount
                properties:
                  keys:
                    $ref: '#/components/schemas/AccountKeys'
                  amounts:
                    type: array
                    items:
                      description: The initial balance of account.
                      allOf:
                        - $ref: '#/components/schemas/Amount'
...
    CurrenciesHAL:
      allOf:
        - $ref: '#/components/schemas/HAL'
        - type: object
          properties:
            _links:
              type: object
              properties:
                self:
                  allOf:
                    - $ref: '#/components/schemas/HALLink'
                    - type: object
                      properties:
                        href:
                          type: string
                          default: /currency/
                          example: /currency/
                currency:{currency_id}:
                  allOf:
                    - $ref: '#/components/schemas/HALLink'
                    - type: object
                      properties:
                        href:
                          type: string
                          default: /currency/XXX
                          example: /currency/XXX
                        templated:
                          type: boolean
                          default: true
                          example: true
    CurrencyHAL:
      allOf:
        - $ref: '#/components/schemas/HAL'
        - type: object
          properties:
             hint:
              type: object
              properties:
                name:
                  type: string
                  default: mitum-currency-currency-design
                  example: mitum-currency-currency-design
                hint:
                  type: string
                  default: a030:0.0.1
                  example: a030:0.0.1
             _embedded:
                $ref: '#/components/schemas/CurrencyDesign'
             _links:
                type: object
                properties:
                  self:
                    allOf:
                      - $ref: '#/components/schemas/HALLink'
                      - type: object
                        properties:
                          href:
                            type: string
                            default: /currency/XXX
                            example: /currency/XXX
    CurrencyID:
      description: currency unique id(or name)
      type: string
      example: XXX
...
    OperationValue:
      type: object
      required:
      - _hint
      - hash
      - operation
      - height
      - confirmed_at
      - in_state
      properties:
        _hint:
          allOf:
            - $ref: '#/components/schemas/Hint'
            - type: string
              example: 0150:0.0.1
              default: 0150:0.0.1
        hash:
          description: operation value hash, not operation hash
          type: string
          format: hash
          example: 6GymxmuvhgUfAKWYQUKKveYh1H8b981vQaZcMsbWqykS
        operation:
          oneOf:
            - $ref: '#/components/schemas/CreateAccounts'
            - $ref: '#/components/schemas/KeyUpdater'
            - $ref: '#/components/schemas/Transfers'
            - $ref: '#/components/schemas/CurrencyRegister'
            - $ref: '#/components/schemas/CurrencyPolicyUpdater'
        height:
          $ref: '#/components/schemas/Height'
        confirmed_at:
          type: string
          format: date-time
          example: "2020-10-13T14:37:20Z"
        in_state:
          description: Whether the operation is in block or not. If `false`, the operation is processed, but ignored with it's own problem.
          type: boolean
          example: true
        reason:
          $ref: '#/components/schemas/ReasonError'
          description: It describes the rejected reason of operation.
    CurrencyDesign:
      type: object
      required:
      - _hint
      - amount
      - genesis_account
      - policy
      properties:
        _hint:
          allOf:
            - $ref: '#/components/schemas/Hint'
            - type: string
              default: a030:0.0.1
              example: a030:0.0.1
        amount:
          $ref: '#/components/schemas/Amount'
        genesis_account:
          allOf:
            - $ref: '#/components/schemas/AccountAddress'
            - description: genesis account address, which will hold genesis balance
        policy:
          $ref: '#/components/schemas/CurrencyPolicy'
    Amount:
      type: object
      required:
      - _hint
      - hash
      - amount
      - currency
      properties:
        _hint:
          allOf:
            - $ref: '#/components/schemas/Hint'
            - type: string
              default: a022:0.0.1
              example: a022:0.0.1
        hash:
          type: string
          format: hash
          example: yaJGkXgizfaB3HEAf9LWdWBC3Sh954pzuc2tNTM9PHG
        amount:
          type: string
          description: amount
          example: 33
          default: 33
        currency:
          $ref: '#/components/schemas/CurrencyID'
    CurrencyPolicy:
      description: currency policy
      type: object
      required:
      - _hint
      - new_account_min_balance
      - feeer
      properties:
        _hint:
          allOf:
            - $ref: '#/components/schemas/Hint'
            - type: string
              default: a036:0.0.1
              example: a036:0.0.1
        _hash:
          type: string
          format: hash
          example: yaJGkXgizfaB3HEAf9LWdWBC3Sh954pzuc2tNTM9PHG
        new_account_min_balance:
          allOf:
            - $ref: '#/components/schemas/Amount'
            - description: minimum balance for new account
        feeer:
          description: fee policy
          type: object
          oneOf:
            - $ref: '#/components/schemas/NilFeeer'
            - $ref: '#/components/schemas/FixedFeeer'
            - $ref: '#/components/schemas/RatioFeeer'
    NilFeeer:
      description: fee policy, which does not charge fee
      type: object
      required:
      - _hint
      properties:
        _hint:
          allOf:
            - $ref: '#/components/schemas/Hint'
            - type: string
              default: a031:0.0.1
              example: a031:0.0.1
        type:
          type: string
          example: 'nil'
          default: 'nil'
...
    AccountAddress:
      description: account address
      type: string
      example: 'mc-node-010a:0.0.1'
    ReasonError:
      type: object
      required:
        - msg
      properties:
        msg:
          description: reason message
          type: string
          example: 'insufficient balance'
        data:
          description: additional data
          type: object

# vi: ft=yaml tw=100 ts=2 sw=2 expandtab smarttab